Merge configuration during writes
At the moment the new configuration is merged only to see if it must
change, but not during the actual file writing.

This difference could cause big diffs when the configuration is
updated from the Mattermost console and it makes hard to understand
what has changed from the default.

This patch makes the generated configuration file more similar to the
one generated by Mattermost itself.